Menu classes refactored
